requirements.txt
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
idna==3.10
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
